+++ CHANGELOG.md
@@ -4,8 +4,13 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.1]
+
+- metrics with all macros
+
 ## [1.6.0]
 
+- can update meal date
 - add confirm on deconnect
 - add prevention message to user that informations are indicative
 - weight as decimal